Conflicts:
	app/assets/stylesheets/application.css.scss
	app/assets/stylesheets/brandscopic.css.scss
	app/assets/stylesheets/font/icons.eot
	app/assets/stylesheets/font/icons.svg
	app/assets/stylesheets/font/icons.ttf
	app/assets/stylesheets/font/icons.woff
	app/assets/stylesheets/icons.css.scss

+++ app/models/membership.rb
@@ -28,23 +28,28 @@ class Membership < ActiveRecord::Base
   validates :memberable_id, presence: true
   validates :memberable_type, presence: true
 
+  validate :same_company
+
   belongs_to :parent, polymorphic: true
 
   private
     def create_notifications
       if memberable_type == 'Campaign' && company_user.role.has_permission?(:read, Campaign)
         Notification.new_campaign(company_user, memberable)
-      elsif memberable_type == 'Event' && company_user.allowed_to_access_place?(memberable.place)
-        Notification.new_event(company_user, memberable)
+      elsif memberable_type == 'Event' &&
+        memberable.company.setting(:event_alerts_policy, Notification::EVENT_ALERT_POLICY_TEAM) == Notification::EVENT_ALERT_POLICY_TEAM &&
+        company_user.allowed_to_access_place?(memberable.place)
+          Notification.new_event(company_user, memberable)
       end
     end
 
     def delete_notifications
       if memberable_type == 'Campaign'
         company_user.notifications.where(path: Rails.application.routes.url_helpers.campaign_path(memberable)).destroy_all
-      elsif memberable_type == 'Event'
-        company_user.notifications.where(path: Rails.application.routes.url_helpers.event_path(memberable)).destroy_all
-        company_user.notifications.where("params->'task_id' in (?)", memberable.task_ids.map{|n| n.to_s}).destroy_all
+      elsif memberable_type == 'Event' &&
+        memberable.company.setting(:event_alerts_policy, Notification::EVENT_ALERT_POLICY_TEAM) == Notification::EVENT_ALERT_POLICY_TEAM
+          company_user.notifications.where(path: Rails.application.routes.url_helpers.event_path(memberable)).destroy_all
+          company_user.notifications.where("params->'task_id' in (?)", memberable.task_ids.map{|n| n.to_s}).destroy_all
       end
     end
 
@@ -63,9 +68,16 @@ class Membership < ActiveRecord::Base
       if memberable.is_a?(Area)
         Rails.cache.delete("user_accessible_locations_#{company_user_id}")
         Rails.cache.delete("user_accessible_places_#{company_user_id}")
-      elsif memberable.is_a?(Campaign)
+      elsif memberable.is_a?(Campaign) || memberable.is_a?(Brand) || memberable.is_a?(BrandPortfolio)
         Rails.cache.delete("user_accessible_campaigns_#{company_user_id}")
       end
       true
+    end
+
+    # Validates that the user and the memberable are from the same company
+    def same_company
+      if company_user.company_id != memberable.company_id
+        errors.add(:memberable_id, :invalid)
+      end
     end
 end
